requirements.txt
# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
 
# Machine Learning and NLP
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
 
# LLM
openai>=1.0.0
 
# PDF processing
PyMuPDF>=1.23.0
 
# Environment variables
python-dotenv>=1.0.0
 
# Data validation
pydantic>=2.0.0
 
# CORS support
fastapi-cors>=0.0.6